requirements.txt
alembic==1.11.1
arrow==1.2.3
black==23.3.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.1.0
cycler==0.11.0
databricks-cli==0.17.7
databricks-connect==13.1.0
databricks-sdk==0.1.11
docker==6.1.3
entrypoints==0.4
Flask==2.3.2
fonttools==4.40.0
gitdb==4.0.10
GitPython==3.1.31
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.56.0
grpcio-status==1.56.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.7.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.2.0
kiwisolver==1.4.4
make==0.1.6.post2
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
mlflow==2.4.1
mypy-extensions==1.0.0
numpy==1.25.0
oauthlib==3.2.2
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.8.0
pluggy==1.2.0
protobuf==4.23.3
py4j==0.10.9.7
pyarrow==12.0.1
pydantic==1.10.9
PyJWT==2.7.0
pyparsing==3.1.0
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
querystring-parser==1.2.4
requests==2.31.0
scikit-learn==1.2.2
scipy==1.11.0
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.17
sqlparse==0.4.4
tabulate==0.9.0
threadpoolctl==3.1.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
websocket-client==1.6.1
Werkzeug==2.3.6
zipp==3.15.0
